Add new method of fetching scenarios using fetchScenariosPost which allows filtering by any item on the scenario during the Post.
Add this methodology into the ScenariosOld component allowing users to input the QueryFilter as a prop which scenarios will then be fetched using.

This allows for more efficient fetching of data as only the scenarios matching the required filters are pulled across, reducing required filtering on the frontend and load on the api.